Replace stub with same method and url
When a stub is added for a method and url that is already stubbed
the existing stub will be replaced by the new one.

This lets you for example set up a set of generic stubs in
`beforeEach` and then override them as needed with more specific
implementations in your testcases.

This could be a breaking change if your testcases previously
depended on newly added stubbs never being called.