stubRequest should accept method
This is a breaking change, but brings the method in line with its own docs, and makes it MUCH more useful.

We could add method as the third argument to avoid the breaking change, but then stubRequest would have a different signature than stubOnce, which is odd. Then again, it already has a different signature - basically its a mess :).

Fixes #5.